refactored modules from dir `helpers`: moved to other dirs, removed
