+++ cache/test_cache.py
@@ -9,10 +9,10 @@ from http import HTTPStatus
 
 from framework.curl_client import CurlResponse
 from framework.deproxy_server import StaticDeproxyServer
-from helpers import checks_for_tests as checks
 from helpers import deproxy
 from helpers.control import Tempesta
 from helpers.deproxy import HttpMessage
+from test_suite import checks_for_tests as checks
 from test_suite import tester
 from test_suite.parameterize import param, parameterize, parameterize_class
