+++ http2_general/test_flow_control_window.py
@@ -7,9 +7,10 @@ __license__ = "GPL2"
 from h2.errors import ErrorCodes
 from hyperframe.frame import DataFrame
 
-from helpers import analyzer, asserts, remote, util
+from helpers import analyzer, remote, util
 from helpers.deproxy import HttpMessage
 from http2_general.helpers import H2Base
+from test_suite import asserts
 
 
 class TestFlowControl(H2Base, asserts.Sniffer):
