+++ selftests/test_x509.py
@@ -2,7 +2,7 @@ import subprocess
 import unittest
 from pathlib import Path
 
-from helpers.x509 import CertGenerator
+from helpers.cert_generator_x509 import CertGenerator
 
 __author__ = "Tempesta Technologies, Inc."
 __copyright__ = "Copyright (C) 2022-2024 Tempesta Technologies, Inc."
