+++ t_long_body/test_long_response.py
@@ -5,10 +5,10 @@ __copyright__ = "Copyright (C) 2022 Tempesta Technologies, Inc."
 __license__ = "GPL2"
 
 from framework import curl_client, deproxy_server
-from helpers import checks_for_tests as checks
 from helpers import deproxy, tf_cfg
 from t_long_body import utils
 from t_stress.test_stress import CustomMtuMixin
+from test_suite import checks_for_tests as checks
 from test_suite.tester import TempestaTest
 
 BODY_SIZE = 1024**2 * int(tf_cfg.cfg.get("General", "long_body_size"))
