+++ t_stress/test_nginx.py
@@ -2,8 +2,8 @@
 HTTP Stress tests with NGINX in place of Tempesta FW - to compare results.
 """
 
+from helpers.cert_generator_x509 import CertGenerator
 from helpers.tf_cfg import cfg
-from helpers.x509 import CertGenerator
 from t_stress.test_stress import BaseCurlStress, BaseWrkStress
 
 __author__ = "Tempesta Technologies, Inc."
