+++ test_suite/custom_error_page.py
@@ -4,6 +4,7 @@ Custom error page generator.
 Tempesta administrator can specify files with page bodies for error responses
 generated by Tempesta FW.
 """
+
 from helpers import tf_cfg
 
 __author__ = "Tempesta Technologies, Inc."
