+++ tls/test_tls_cert.py
@@ -11,9 +11,9 @@ from itertools import cycle, islice
 from cryptography.hazmat.primitives.asymmetric import ec
 
 from helpers import dmesg, remote, tempesta, tf_cfg
+from helpers.cert_generator_x509 import CertGenerator
 from helpers.deproxy import HttpMessage
 from helpers.util import fill_template
-from helpers.x509 import CertGenerator
 from test_suite import tester
 
 from .handshake import TlsHandshake, x509_check_cn
